requirements.txt
featurecloud
protobuf==3.20.*
tensorflow==2.2.0
numpy==1.19.2
PyYAML
scipy
pandas
scikit-learn
matplotlib
